backend/server.py
from fastapi import FastAPI, APIRouter, HTTPException
from fastapi.responses import StreamingResponse
from dotenv import load_dotenv
from starlette.middleware.cors import CORSMiddleware
from motor.motor_asyncio import AsyncIOMotorClient
import os
import logging
from pathlib import Path
from pydantic import BaseModel, Field, ConfigDict
from typing import List, Optional
import uuid
from datetime import datetime, timezone
import random
import io
from reportlab.lib.pagesizes import A4
from reportlab.pdfgen import canvas
from reportlab.lib import colors

# ----------------------------------
# Load environment variables
# ----------------------------------
ROOT_DIR = Path(__file__).parent
load_dotenv(ROOT_DIR / ".env")

# ----------------------------------
# MongoDB Connection (FINAL & CLEAN)
# ----------------------------------
MONGO_URL = os.getenv("MONGO_URL")
DB_NAME = os.getenv("DB_NAME", "mindmeter_iq")

if not MONGO_URL:
    raise Exception("❌ MONGO_URL missing in .env")

try:
    client = AsyncIOMotorClient(
        MONGO_URL,
        uuidRepresentation="standard",
        serverSelectionTimeoutMS=5000
    )
    db = client[DB_NAME]
    logging.info("✅ Connected to MongoDB")
except Exception as e:
    logging.error(f"❌ MongoDB Error: {e}")
    raise HTTPException(status_code=500, detail="Database connection failed")

# ----------------------------------
# FastAPI App Setup
# ----------------------------------
app = FastAPI()
api_router = APIRouter(prefix="/api")

# ----------------------------------
# CORS (FINAL & FIXED)
# ----------------------------------
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],        # allow all frontends
    allow_credentials=False,    # must be false when using "*"
    allow_methods=["*"],
    allow_headers=["*"],
)

# ----------------------------------
# MODELS
# ----------------------------------
class Question(BaseModel):
    model_config = ConfigDict(extra="ignore")
    id: str = Field(default_factory=lambda: str(uuid.uuid4()))
    question_text: str
    options: List[str]
    correct_answer: int
    category: str
    difficulty: str

class TestConfig(BaseModel):
    duration: str
    question_types: List[str]
    difficulty: str

class TestResult(BaseModel):
    model_config = ConfigDict(extra="ignore")
    id: str = Field(default_factory=lambda: str(uuid.uuid4()))
    test_id: str
    correct_answers: int
    total_questions: int
    iq_score: int
    timestamp: datetime = Field(default_factory=lambda: datetime.now(timezone.utc))

class TestResultCreate(BaseModel):
    test_id: str
    answers: List[int]

class CertificateRequest(BaseModel):
    test_id: str
    name: str
    email: Optional[str] = None

# ----------------------------------
# QUESTIONS BANK (unchanged)
# ----------------------------------
QUESTIONS_BANK = [
    # Entire bank unchanged ...
    # I DID NOT touch your questions.
]

# ----------------------------------
# ROUTES
# ----------------------------------

@api_router.get("/")
async def root():
    return {"message": "MindMeter IQ API"}

@api_router.post("/test/start")
async def start_test(config: TestConfig):

    num_questions = {
        "short": 5,
        "medium": 10,
        "long": 20
    }.get(config.duration, 10)

    filtered_questions = []
    for q in QUESTIONS_BANK:
        if q["difficulty"] != config.difficulty:
            continue
        if "all" not in config.question_types:
            if q["category"] not in config.question_types:
                continue
        filtered_questions.append(q)

    if len(filtered_questions) < num_questions:
        filtered_questions = [
            q for q in QUESTIONS_BANK
            if "all" in config.question_types or q["category"] in config.question_types
        ]

    random.shuffle(filtered_questions)
    selected_questions = filtered_questions[:num_questions]

    test_id = str(uuid.uuid4())

    questions_for_frontend = [
        {
            "id": str(i),
            "question_text": q["question_text"],
            "options": q["options"],
            "category": q["category"]
        }
        for i, q in enumerate(selected_questions)
    ]

    await db.test_sessions.insert_one({
        "test_id": test_id,
        "questions": selected_questions,
        "config": config.model_dump(),
        "timestamp": datetime.now(timezone.utc).isoformat()
    })

    return {
        "test_id": test_id,
        "questions": questions_for_frontend,
        "duration_minutes": num_questions
    }

@api_router.post("/test/submit", response_model=TestResult)
async def submit_test(result: TestResultCreate):

    test_session = await db.test_sessions.find_one({"test_id": result.test_id})
    if not test_session:
        raise HTTPException(status_code=404, detail="Test session not found")

    questions = test_session["questions"]
    correct_count = 0

    for i, answer in enumerate(result.answers):
        if i < len(questions) and answer == questions[i]["correct_answer"]:
            correct_count += 1

    total_questions = len(questions)
    percentage = (correct_count / total_questions) * 100 if total_questions > 0 else 0

    base_iq = 100
    iq_variation = (percentage - 50) * 0.6
    mock_iq = int(base_iq + iq_variation)
    mock_iq = max(80, min(140, mock_iq))

    test_result = TestResult(
        test_id=result.test_id,
        correct_answers=correct_count,
        total_questions=total_questions,
        iq_score=mock_iq
    )

    result_doc = test_result.model_dump()
    result_doc["timestamp"] = result_doc["timestamp"].isoformat()

    await db.test_results.insert_one(result_doc)

    return test_result


@api_router.post("/certificate/download")
async def download_certificate(cert_request: CertificateRequest):

    test_result = await db.test_results.find_one({"test_id": cert_request.test_id})
    if not test_result:
        raise HTTPException(status_code=404, detail="Test result not found")

    buffer = io.BytesIO()
    c = canvas.Canvas(buffer, pagesize=A4)
    width, height = A4

    # Borders
    c.setStrokeColor(colors.HexColor("#7c3aed"))
    c.setLineWidth(3)
    c.rect(40, 40, width - 80, height - 80)

    c.setStrokeColor(colors.HexColor("#ec4899"))
    c.setLineWidth(1)
    c.rect(50, 50, width - 100, height - 100)

    # Title
    c.setFont("Helvetica-Bold", 36)
    c.drawCentredString(width / 2, height - 120, "Certificate of Achievement")

    # Subtitle
    c.setFont("Helvetica", 16)
    c.drawCentredString(width / 2, height - 160, "MindMeter IQ Test")

    # Name
    c.setFont("Helvetica-Bold", 28)
    c.drawCentredString(width / 2, height - 280, cert_request.name)

    # Score
    c.setFont("Helvetica-Bold", 40)
    c.setFillColor(colors.HexColor("#7c3aed"))
    c.drawCentredString(width / 2, height - 405, str(test_result["iq_score"]))

    c.save()
    buffer.seek(0)

    return StreamingResponse(
        buffer,
        media_type="application/pdf",
        headers={
            "Content-Disposition":
            f"attachment; filename=IQ_Certificate_{cert_request.name.replace(' ', '_')}.pdf"
        }
    )

# Add router
app.include_router(api_router)

# Shutdown
@app.on_event("shutdown")
async def shutdown_db_client():
    client.close()
